# One way allocator to double the speed of parallel context queries