--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -383,6 +383,8 @@ set(LIBNETDATA_FILES
         libnetdata/log/log.h
         libnetdata/os.c
         libnetdata/os.h
+        libnetdata/onewayalloc/onewayalloc.c
+        libnetdata/onewayalloc/onewayalloc.h
         libnetdata/popen/popen.c
         libnetdata/popen/popen.h
         libnetdata/procfile/procfile.c