--- a/configure.ac
+++ b/configure.ac
@@ -1769,6 +1769,7 @@ AC_CONFIG_FILES([
     libnetdata/eval/Makefile
     libnetdata/locks/Makefile
     libnetdata/log/Makefile
+    libnetdata/onewayalloc/Makefile
     libnetdata/popen/Makefile
     libnetdata/procfile/Makefile
     libnetdata/simple_pattern/Makefile