--- a/libnetdata/libnetdata.h
+++ b/libnetdata/libnetdata.h
@@ -342,6 +342,7 @@ extern char *netdata_configured_host_prefix;
 #include "json/json.h"
 #include "health/health.h"
 #include "string/utf8.h"
+#include "onewayalloc/onewayalloc.h"
 
 // BEWARE: Outside of the C code this also exists in alarm-notify.sh
 #define DEFAULT_CLOUD_BASE_URL "https://app.netdata.cloud"