--- a/libnetdata/onewayalloc/onewayalloc.h
+++ b/libnetdata/onewayalloc/onewayalloc.h
@@ -12,5 +12,6 @@ extern void *onewayalloc_mallocz(ONEWAYALLOC *owa, size_t size);
 extern void *onewayalloc_callocz(ONEWAYALLOC *owa, size_t nmemb, size_t size);
 extern char *onewayalloc_strdupz(ONEWAYALLOC *owa, const char *s);
 extern void *onewayalloc_memdupz(ONEWAYALLOC *owa, const void *src, size_t size);
+extern void onewayalloc_freez(ONEWAYALLOC *owa, const void *ptr);
 
 #endif // ONEWAYALLOC_H