Context queries do a gazillion of `mallocz()`, `strdupz()` and `freez()` to create the virtual chart that takes all the dimensions of the context.

This PR adds a `one way allocator`, a set of functions that allocate large blocks using `mmap()` (the fastest possible) and then calls to `onewayalloc_mallocz()` returns blocks of memory within the mmap allocated buffer.

It allocates more and more mmap buffers as needed, transparently, but it takes into account the amount of memory already allocated, to minimize future allocations.

- Allocation are done using `mmap()` with flags: `MAP_ANOMYMOYS|MAP_PRIVATE`. KSM is disabled for them, since this is expected to be short lived
- Allocations are aligned to the CPU `sizeof(size_t)` size. So, the minimum allocation is for that size and all the allocations are always multiple of that. However the caller may request whatever size (the alignment is transparent).
- No locks and no threading support - use this only when you need a super fast allocator for a task
- No `realloc()` support - just allocate a new one
- No `free()` support to deallocate the individual allocations made with this
- To release the memory, destroy the whole OWA when you finish

## How much less allocations?

Tested with a context query, on 1000 charts, having 100 dimensions each (go.d.plugin, `example` module). The result is this: 

```
OWA: 400001 allocations of 60069032 total bytes, in 22 pages of 63832064 total bytes
```

So, for 100k dimensions total (1000 charts x 100 dimensions each), the code is performing 400k allocations of 60MB total. The OWA allocator did only 22 `mmap()` allocations, of 63MB (a little more, due to the forward looking optimizer).

The reduction of system calls is astonishing (400k vs 22)!

## How faster it is?

On modern hardware, for any single query the performance improvement is marginal and not noticeable at all.

We performed the following tests using the same huge context query (1000 charts, 100 dimensions each = 100k dimensions)

1. current `master` branch, 1 caller, 256 queries (sequential)
2. current `master` branch, 256 callers, 1 query each (parallel)
3. This PR, 1 caller, 256 queries (sequential)
4. This PR, 256 callers, 1 query each (parallel)

Netdata was configured to use 24 web threads on the 24 core server we used.

The results are as follows:

### sequential test

branch|transactions|time to complete|transaction rate|average response time|min response time|max response time
:---:|:---:|:---:|:---:|:---:|:---:|:---:|
`master`|256|322.35s|0.79/sec|1.26s|1.01s|1.87s
this PR|256|310.19s|0.83/sec|1.21s|1.04s|1.63s

For a single query, the improvement is just marginal and not noticeable at all.

### parallel test

branch|transactions|time to complete|transaction rate|average response time|min response time|max response time
:---:|:---:|:---:|:---:|:---:|:---:|:---:|
`master`|256|84.72s|3.02/sec|68.43s|50.20s|84.71s
this PR|256|39.35s|6.51/sec|34.48s|20.55s|39.34s

This PR provides a 54% overall speed improvement (more than double the overall user-experienced speed, including the data query itself).

---

## TODO

The query engine that uses this PR is currently messy. The SQLite parts of it have not been updated to use OWA. So, if this is used in production and the SQLite code kicks in, there will be memory leaks (the code that frees malloc() allocations is commented, since it is not supported by OWA.

So, to merge this PR, the sqlite code needs to be updated too.